Move functions from header to cpp (#111)
* Move functions from header to cpp

* Move all non-templated code to cpp files

* UPdate wrapper
