+++ cpp/CMakeLists.txt
@@ -53,7 +53,7 @@ target_link_libraries(dolfinx_contact PUBLIC Basix::basix)
 target_link_libraries(dolfinx_contact PUBLIC dolfinx)
 
 include(GNUInstallDirs)
-install(FILES Contact.h contact_kernels.hpp error_handling.h utils.h coefficients.h elasticity.h geometric_quantities.h meshtie_kernels.h SubMesh.h QuadratureRule.h RayTracing.h KernelData.h DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/dolfinx_contact COMPONENT Development)
+install(FILES Contact.h contact_kernels.h error_handling.h utils.h coefficients.h elasticity.h geometric_quantities.h meshtie_kernels.h SubMesh.h QuadratureRule.h RayTracing.h KernelData.h DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/dolfinx_contact COMPONENT Development)
 
 target_sources(dolfinx_contact PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/utils.cpp
@@ -67,6 +67,7 @@ target_sources(dolfinx_contact PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/RayTracing.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/KernelData.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/error_handling.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/contact_kernels.cpp
   )
 
 # Set target include location (for build and installed)
