+++ cpp/Contact.h
@@ -31,31 +31,6 @@ using mat_set_fn = const std::function<int(
 namespace dolfinx_contact
 {
 
-namespace impl
-{
-/// Tabulate the coordinate element basis functions at quadrature points
-///
-/// @param[in] cmap The coordinate element
-/// @param[in] q_rule The quadrature rule
-std::pair<std::vector<double>, std::array<std::size_t, 4>>
-tabulate(const dolfinx::fem::CoordinateElement& cmap,
-         std::shared_ptr<const dolfinx_contact::QuadratureRule> q_rule)
-{
-
-  // Create quadrature points on reference facet
-  const std::vector<double>& q_weights = q_rule->weights();
-  const std::vector<double>& q_points = q_rule->points();
-  assert(q_weights.size() == (std::size_t)q_rule->offset().back());
-  // Tabulate Coordinate element (first derivative to compute Jacobian)
-  std::array<std::size_t, 4> cmap_shape
-      = cmap.tabulate_shape(0, q_weights.size());
-  std::vector<double> cmap_basis(
-      std::reduce(cmap_shape.begin(), cmap_shape.end(), 1, std::multiplies{}));
-  cmap.tabulate(0, q_points, {q_weights.size(), q_rule->tdim()}, cmap_basis);
-  return {cmap_basis, cmap_shape};
-}
-} // namespace impl
-
 class Contact
 {
 public:
@@ -191,61 +166,13 @@ public:
   /// physical facet for each facet in _facet_"origin_meshtag" Creates and
   /// fills _qp_phys_"origin_meshtag"
   /// @param[in] origin_meshtag flag to choose the surface
-  void create_q_phys(int origin_meshtag)
-  {
-    // Get information depending on surface
-    const SubMesh& submesh = _submeshes[origin_meshtag];
-
-    const std::vector<std::int32_t> submesh_facets
-        = submesh.get_submesh_tuples(_cell_facet_pairs[origin_meshtag]);
-    auto mesh_sub = submesh.mesh();
-    const std::size_t gdim = mesh_sub->geometry().dim();
-    const std::vector<size_t>& qp_offsets = _quadrature_rule->offset();
-    _qp_phys[origin_meshtag].resize((qp_offsets[1] - qp_offsets[0])
-                                    * (submesh_facets.size() / 2) * gdim);
-    compute_physical_points(
-        *mesh_sub, submesh_facets, qp_offsets,
-        cmdspan4_t(_reference_basis.data(), _reference_shape),
-        _qp_phys[origin_meshtag]);
-  }
+  void create_q_phys(int origin_meshtag);
 
   /// Compute maximum number of links
   /// I think this should actually be part of create_distance_map
   /// which should be easier after the rewrite of contact
   /// It is therefore called inside create_distance_map
-  void max_links(int pair)
-  {
-    std::size_t max_links = 0;
-    // Select which side of the contact interface to loop from and get the
-    // correct map
-    const std::array<int, 2>& contact_pair = _contact_pairs[pair];
-    const std::vector<std::int32_t>& active_facets
-        = _cell_facet_pairs[contact_pair.front()];
-    std::shared_ptr<const dolfinx::graph::AdjacencyList<int>> map
-        = _facet_maps[pair];
-    assert(map);
-    std::shared_ptr<const dolfinx::graph::AdjacencyList<int>> facet_map
-        = _submeshes[contact_pair.back()].facet_map();
-    assert(facet_map);
-    for (std::size_t i = 0; i < active_facets.size(); i += 2)
-    {
-      std::vector<std::int32_t> linked_cells;
-      for (auto link : map->links((int)i / 2))
-      {
-        if (link >= 0)
-        {
-          auto facet_pair = facet_map->links(link);
-          linked_cells.push_back(facet_pair.front());
-        }
-      }
-      // Remove duplicates
-      std::sort(linked_cells.begin(), linked_cells.end());
-      linked_cells.erase(std::unique(linked_cells.begin(), linked_cells.end()),
-                         linked_cells.end());
-      max_links = std::max(max_links, linked_cells.size());
-    }
-    _max_links[pair] = max_links;
-  }
+  void max_links(int pair);
 
   /// For a given contact pair, for quadrature point on the first surface
   /// compute the closest candidate facet on the second surface.
@@ -262,232 +189,15 @@ public:
   /// @param[out] c - gap packed on facets. c[i*cstride +  gdim * k+ j]
   /// contains the jth component of the Gap on the ith facet at kth
   /// quadrature point
-  std::pair<std::vector<PetscScalar>, int> pack_gap(int pair)
-  {
-    // FIXME: This function should take in the quadrature points
-    // (push_forward_quadrature) of the relevant facet, and the reference
-    // points on the other surface (output of distance map)
-    auto [quadrature_mt, candidate_mt] = _contact_pairs[pair];
-
-    const std::shared_ptr<const dolfinx::mesh::Mesh>& quadrature_mesh
-        = _submeshes[quadrature_mt].mesh();
-    assert(quadrature_mesh);
-    // Get (cell, local_facet_index) tuples on quadrature submesh
-    const std::vector<std::int32_t> quadrature_facets
-        = _submeshes[quadrature_mt].get_submesh_tuples(
-            _cell_facet_pairs[quadrature_mt]);
-
-    const std::shared_ptr<const dolfinx::mesh::Mesh>& candidate_mesh
-        = _submeshes[candidate_mt].mesh();
-    assert(candidate_mesh);
-    // Get (cell, local_facet_index) tuples on candidate submesh
-    const std::vector<std::int32_t> candidate_facets
-        = _submeshes[candidate_mt].get_submesh_tuples(
-            _cell_facet_pairs[candidate_mt]);
-
-    auto [candidate_map, reference_x, shape] = compute_distance_map(
-        *quadrature_mesh, quadrature_facets, *candidate_mesh, candidate_facets,
-        *_quadrature_rule, _mode);
-
-    // NOTE: Assumes same number of quadrature points on all facets
-    error::check_cell_type(candidate_mesh->topology().cell_type());
-    const std::size_t num_q_point
-        = _quadrature_rule->offset()[1] - _quadrature_rule->offset()[0];
-    const std::size_t num_facets = _cell_facet_pairs[quadrature_mt].size() / 2;
-    const int q_gdim = quadrature_mesh->geometry().dim();
-    mdspan3_t qp_span(_qp_phys[quadrature_mt].data(), num_facets, num_q_point,
-                      q_gdim);
-
-    // Get information about submesh geometry and topology
-    const dolfinx::mesh::Geometry& geometry = candidate_mesh->geometry();
-    const int gdim = geometry.dim();
-    std::span<const double> x_g = geometry.x();
-    auto x_dofmap = geometry.dofmap();
-    const dolfinx::fem::CoordinateElement& cmap = geometry.cmap();
-    const std::size_t num_dofs_g = cmap.dim();
-    const dolfinx::mesh::Topology& topology = candidate_mesh->topology();
-    const int tdim = topology.dim();
-
-    std::vector<double> coordinate_dofsb(num_dofs_g * gdim);
-    cmdspan2_t coordinate_dofs(coordinate_dofsb.data(), num_dofs_g, gdim);
-    std::array<double, 3> coordb;
-    mdspan2_t coord(coordb.data(), 1, gdim);
-
-    // Pack gap function for each quadrature point on each facet
-    std::vector<PetscScalar> c(num_facets * num_q_point * gdim, 0.0);
-    const int cstride = (int)num_q_point * gdim;
-
-    auto f_to_c = candidate_mesh->topology().connectivity(tdim - 1, tdim);
-    if (!f_to_c)
-    {
-      throw std::runtime_error("Missing facet to cell connectivity on "
-                               "candidate submesh");
-    }
-    const std::array<std::size_t, 4> basis_shape
-        = cmap.tabulate_shape(0, shape[0]);
-    assert(basis_shape.back() == 1);
-    std::vector<double> cmap_basis(std::reduce(
-        basis_shape.begin(), basis_shape.end(), 1, std::multiplies{}));
-    cmap.tabulate(0, reference_x, shape, cmap_basis);
-    cmdspan4_t full_basis(cmap_basis.data(), basis_shape);
-    for (std::size_t i = 0; i < num_facets; ++i)
-    {
-      int offset = (int)i * cstride;
-      auto facets = candidate_map.links(i);
-      assert(facets.size() == num_q_point);
-
-      for (std::size_t q = 0; q < num_q_point; ++q)
-      {
-
-        // Skip negative facet indices (No facet on opposite surface has
-        // been found)
-        if (facets[q] < 0)
-          continue;
-
-        auto candidate_cells = f_to_c->links(facets[q]);
-        assert(candidate_cells.size() == 1);
-
-        // Copy coordinate dofs of candidate cell
-        // Get cell geometry (coordinate dofs)
-        auto x_dofs = x_dofmap.links(candidate_cells.front());
-        assert(x_dofs.size() == num_dofs_g);
-        for (std::size_t j = 0; j < num_dofs_g; ++j)
-        {
-          std::copy_n(std::next(x_g.begin(), 3 * x_dofs[j]), gdim,
-                      std::next(coordinate_dofsb.begin(), j * gdim));
-        }
-
-        auto basis_q = stdex::submdspan(
-            full_basis, 0,
-            std::pair{i * num_q_point + q, i * num_q_point + q + 1},
-            stdex::full_extent, 0);
-
-        dolfinx::fem::CoordinateElement::push_forward(coord, coordinate_dofs,
-                                                      basis_q);
-        for (int k = 0; k < gdim; k++)
-          c[offset + q * gdim + k] = coordb[k] - qp_span(i, q, k);
-      }
-    }
-    return {std::move(c), cstride};
-  }
+  std::pair<std::vector<PetscScalar>, int> pack_gap(int pair);
 
   /// Compute test functions on opposite surface at quadrature points of
   /// facets
   /// @param[in] pair - index of contact pair
   /// @param[in] gap - gap packed on facets per quadrature point
   /// @param[out] c - test functions packed on facets.
-  std::pair<std::vector<PetscScalar>, int> pack_test_functions(int pair)
-  {
-    auto [quadrature_mt, candidate_mt] = _contact_pairs[pair];
+  std::pair<std::vector<PetscScalar>, int> pack_test_functions(int pair);
 
-    // Get mesh info for candidate side
-    const std::shared_ptr<const dolfinx::mesh::Mesh>& candidate_mesh
-        = _submeshes[candidate_mt].mesh();
-    assert(candidate_mesh);
-    const std::shared_ptr<const dolfinx::mesh::Mesh>& quadrature_mesh
-        = _submeshes[quadrature_mt].mesh();
-    assert(quadrature_mesh);
-
-    // Get (cell, local_facet_index) tuples on quadrature submesh
-    const std::vector<std::int32_t> quadrature_facets
-        = _submeshes[quadrature_mt].get_submesh_tuples(
-            _cell_facet_pairs[quadrature_mt]);
-
-    // Get (cell, local_facet_index) tuples on candidate submesh
-    const std::vector<std::int32_t> candidate_facets
-        = _submeshes[candidate_mt].get_submesh_tuples(
-            _cell_facet_pairs[candidate_mt]);
-
-    auto [candidate_map, reference_x, shape]
-        = dolfinx_contact::compute_distance_map(
-            *quadrature_mesh, quadrature_facets, *candidate_mesh,
-            candidate_facets, *_quadrature_rule, _mode);
-
-    // Compute values of basis functions for all y = Pi(x) in qp
-    auto V_sub = std::make_shared<dolfinx::fem::FunctionSpace>(
-        _submeshes[candidate_mt].create_functionspace(_V));
-
-    std::shared_ptr<const dolfinx::fem::FiniteElement> element
-        = V_sub->element();
-    std::array<std::size_t, 4> b_shape
-        = element->basix_element().tabulate_shape(0, shape[0]);
-    if (b_shape.back() > 1)
-      throw std::invalid_argument("pack_test_functions assumes values size 1");
-    std::vector<double> basis_valuesb(
-        std::reduce(b_shape.cbegin(), b_shape.cend(), 1, std::multiplies{}));
-    element->tabulate(basis_valuesb, reference_x, shape, 0);
-    cmdspan4_t basis_values(basis_valuesb.data(), b_shape);
-
-    // Need to apply push forward and dof transformations to test functions
-    assert((b_shape.front() == 1) and (b_shape.back() == 1));
-
-    const basix::FiniteElement& b_el = element->basix_element();
-    if (element->needs_dof_transformations()
-        or b_el.map_type() != basix::maps::type::identity)
-    {
-      // If we want to do this we need to apply transformation and push
-      // forward
-      throw std::runtime_error(
-          "Packing basis (test) functions of space that uses "
-          "non-indentity maps is not supported");
-    }
-
-    // Convert facet index on candidate mesh into cell index
-    const dolfinx::mesh::Topology& topology = candidate_mesh->topology();
-    const int tdim = topology.dim();
-    auto f_to_c = topology.connectivity(tdim - 1, tdim);
-    assert(f_to_c);
-    const std::vector<std::int32_t>& facets = candidate_map.array();
-    error::check_cell_type(topology.cell_type());
-    const std::size_t num_q_points
-        = _quadrature_rule->offset()[1] - _quadrature_rule->offset()[0];
-    const std::size_t num_facets = quadrature_facets.size() / 2;
-    assert(num_facets * num_q_points == facets.size());
-    std::vector<std::int32_t> cells(facets.size(), -1);
-    for (std::size_t i = 0; i < cells.size(); ++i)
-    {
-      if (facets[i] < 0)
-        continue;
-      auto f_cells = f_to_c->links(facets[i]);
-      assert(f_cells.size() == 1);
-      cells[i] = f_cells.front();
-    }
-    // FIXME: Aim to remove this as it depends on the state of the contact
-    // algorithm
-    const std::size_t max_links
-        = *std::max_element(_max_links.begin(), _max_links.end());
-
-    const std::size_t bs = element->block_size();
-    const auto cstride = int(num_q_points * max_links * b_shape[2] * bs);
-    std::vector<PetscScalar> cb(
-        num_facets * max_links * num_q_points * b_shape[2] * bs, 0.0);
-    stdex::mdspan<PetscScalar, stdex::dextents<std::size_t, 5>> c(
-        cb.data(), num_facets, max_links, b_shape[2], num_q_points, bs);
-
-    std::vector<std::int32_t> perm(num_q_points);
-    for (std::size_t i = 0; i < c.extent(0); ++i)
-    {
-      std::span<const std::int32_t> f_cells(cells.data() + i * num_q_points,
-                                            num_q_points);
-      auto [unique_cells, offsets] = sort_cells(f_cells, perm);
-      for (std::size_t j = 0; j < unique_cells.size(); ++j)
-      {
-        auto indices
-            = std::span(perm.data() + offsets[j], offsets[j + 1] - offsets[j]);
-
-        assert(perm.size() >= (std::size_t)offsets[j + 1]);
-        for (std::size_t k = 0; k < c.extent(2); ++k)
-          for (std::size_t q = 0; q < indices.size(); ++q)
-            for (std::size_t l = 0; l < c.extent(4); ++l)
-            {
-              c(i, j, k, indices[q], l)
-                  = basis_values(0, i * num_q_points + indices[q], k, 0);
-            }
-      }
-    }
-
-    return {std::move(cb), cstride};
-  }
   /// Compute gradient of test functions on opposite surface (initial
   /// configuration) at quadrature points of facets
   /// @param[in] pair - index of contact pair
@@ -505,131 +215,7 @@ public:
   /// @param[out] c - test functions packed on facets.
   std::pair<std::vector<PetscScalar>, int>
   pack_u_contact(int pair,
-                 std::shared_ptr<dolfinx::fem::Function<PetscScalar>> u)
-  {
-    dolfinx::common::Timer t("Pack contact u");
-    auto [quadrature_mt, candidate_mt] = _contact_pairs[pair];
-
-    // Get mesh info for candidate side
-    const std::shared_ptr<const dolfinx::mesh::Mesh>& candidate_mesh
-        = _submeshes[candidate_mt].mesh();
-    assert(candidate_mesh);
-    const std::shared_ptr<const dolfinx::mesh::Mesh>& quadrature_mesh
-        = _submeshes[quadrature_mt].mesh();
-    assert(quadrature_mesh);
-
-    // Get (cell, local_facet_index) tuples on quadrature submesh
-    const std::vector<std::int32_t> quadrature_facets
-        = _submeshes[quadrature_mt].get_submesh_tuples(
-            _cell_facet_pairs[quadrature_mt]);
-
-    // Get (cell, local_facet_index) tuples on candidate submesh
-    const std::vector<std::int32_t> candidate_facets
-        = _submeshes[candidate_mt].get_submesh_tuples(
-            _cell_facet_pairs[candidate_mt]);
-
-    auto [candidate_map, reference_x, shape]
-        = dolfinx_contact::compute_distance_map(
-            *quadrature_mesh, quadrature_facets, *candidate_mesh,
-            candidate_facets, *_quadrature_rule, _mode);
-
-    // Compute values of basis functions for all y = Pi(x) in qp
-    auto V_sub = std::make_shared<dolfinx::fem::FunctionSpace>(
-        _submeshes[candidate_mt].create_functionspace(_V));
-    dolfinx::fem::Function<PetscScalar> u_sub(V_sub);
-    std::shared_ptr<const dolfinx::fem::DofMap> sub_dofmap = V_sub->dofmap();
-    assert(sub_dofmap);
-    const int bs_dof = sub_dofmap->bs();
-    _submeshes[candidate_mt].copy_function(*u, u_sub);
-
-    std::shared_ptr<const dolfinx::fem::FiniteElement> element
-        = V_sub->element();
-    std::array<std::size_t, 4> b_shape
-        = element->basix_element().tabulate_shape(0, shape[0]);
-    if (b_shape.back() > 1)
-      throw std::invalid_argument("pack_test_functions assumes values size 1");
-    std::vector<double> basis_valuesb(
-        std::reduce(b_shape.cbegin(), b_shape.cend(), 1, std::multiplies{}));
-    element->tabulate(basis_valuesb, reference_x, shape, 0);
-
-    // Need to apply push forward and dof transformations to test functions
-    assert((b_shape.front() == 1) and (b_shape.back() == 1));
-
-    const basix::FiniteElement& b_el = element->basix_element();
-    if (element->needs_dof_transformations()
-        or b_el.map_type() != basix::maps::type::identity)
-    {
-      // If we want to do this we need to apply transformation and push
-      // forward
-      throw std::runtime_error(
-          "Packing u on opposite surface functions of space that uses "
-          "non-indentity maps is not supported");
-    }
-
-    cmdspan4_t basis_values(basis_valuesb.data(), b_shape);
-    const std::span<const PetscScalar>& u_coeffs = u_sub.x()->array();
-
-    // Output vector
-    const std::size_t num_facets = quadrature_facets.size() / 2;
-    const dolfinx::mesh::Topology& topology = candidate_mesh->topology();
-    error::check_cell_type(topology.cell_type());
-    const std::size_t bs_element = element->block_size();
-
-    const std::size_t num_q_points
-        = _quadrature_rule->offset()[1] - _quadrature_rule->offset()[0];
-    std::vector<PetscScalar> c(num_facets * num_q_points * bs_element, 0.0);
-
-    // Get cell index on sub-mesh
-    const int tdim = topology.dim();
-    auto f_to_c = topology.connectivity(tdim - 1, tdim);
-    assert(f_to_c);
-    const std::vector<std::int32_t>& facets = candidate_map.array();
-    std::vector<std::int32_t> cells(facets.size(), -1);
-    for (std::size_t i = 0; i < cells.size(); ++i)
-    {
-      if (facets[i] < 0)
-        continue;
-      auto f_cells = f_to_c->links(facets[i]);
-      assert(f_cells.size() == 1);
-      cells[i] = f_cells.front();
-    }
-
-    // Create work vector for expansion coefficients
-    const auto cstride = int(num_q_points * bs_element);
-    const std::size_t num_basis_functions = b_shape[2];
-    const std::size_t value_size = b_shape[3];
-    std::vector<PetscScalar> coefficients(num_basis_functions * bs_element);
-
-    for (std::size_t i = 0; i < num_facets; ++i)
-    {
-      for (std::size_t q = 0; q < num_q_points; ++q)
-      {
-        // Get degrees of freedom for current cell
-        if (facets[i * num_q_points + q] < 0)
-          continue;
-        auto dofs = sub_dofmap->cell_dofs(cells[i * num_q_points + q]);
-        for (std::size_t j = 0; j < dofs.size(); ++j)
-          for (int k = 0; k < bs_dof; ++k)
-            coefficients[bs_dof * j + k] = u_coeffs[bs_dof * dofs[j] + k];
-
-        // Compute expansion
-        for (std::size_t k = 0; k < bs_element; ++k)
-        {
-          for (std::size_t l = 0; l < num_basis_functions; ++l)
-          {
-            for (std::size_t m = 0; m < value_size; ++m)
-            {
-              c[cstride * i + q * bs_element + k]
-                  += coefficients[bs_element * l + k]
-                     * basis_values(0, num_q_points * i + q, l, m);
-            }
-          }
-        }
-      }
-    }
-    t.stop();
-    return {std::move(c), cstride};
-  }
+                 std::shared_ptr<dolfinx::fem::Function<PetscScalar>> u);
 
   /// Compute gradient of function on opposite surface at quadrature points of
   /// facets
@@ -660,38 +246,8 @@ public:
   /// @param[in] g - defines location of plane
   /// @param[out] c - gap packed on facets. c[i, gdim * k+ j] contains the
   /// jth component of the Gap on the ith facet at kth quadrature point
-  std::pair<std::vector<PetscScalar>, int> pack_gap_plane(int pair, double g)
-  {
-    int quadrature_mt = _contact_pairs[pair][0];
-    // Mesh info
-    std::shared_ptr<const dolfinx::mesh::Mesh> mesh = _V->mesh();
-    assert(mesh);
+  std::pair<std::vector<PetscScalar>, int> pack_gap_plane(int pair, double g);
 
-    const int gdim = mesh->geometry().dim(); // geometrical dimension
-
-    // Tabulate basis function on reference cell (_phi_ref_facets)
-    const dolfinx::fem::CoordinateElement& cmap = mesh->geometry().cmap();
-    std::tie(_reference_basis, _reference_shape)
-        = impl::tabulate(cmap, _quadrature_rule);
-
-    // Compute quadrature points on physical facet _qp_phys_"quadrature_mt"
-    create_q_phys(quadrature_mt);
-    const std::size_t num_facets = _cell_facet_pairs[quadrature_mt].size() / 2;
-    // FIXME: This does not work for prism meshes
-    std::size_t num_q_point
-        = _quadrature_rule->offset()[1] - _quadrature_rule->offset()[0];
-    mdspan3_t qp_span(_qp_phys[quadrature_mt].data(), num_facets, num_q_point,
-                      gdim);
-    std::vector<PetscScalar> c(num_facets * num_q_point * gdim, 0.0);
-    const int cstride = (int)num_q_point * gdim;
-    for (std::size_t i = 0; i < num_facets; i++)
-    {
-      int offset = (int)i * cstride;
-      for (std::size_t k = 0; k < num_q_point; k++)
-        c[offset + (k + 1) * gdim - 1] = g - qp_span(i, k, gdim - 1);
-    }
-    return {std::move(c), cstride};
-  }
   /// This function updates the submesh geometry for all submeshes using
   /// a function given on the parent mesh
   /// @param[in] u - displacement
